services/iris_detection/requirements.txt
numpy==1.23.2
easydict
coremltools
protobuf==3.20.1
mediapipe
Pillow